Django/CodingTestScoringServer/scoringserver/requirements.txt
asgiref==3.5.2
certifi==2023.5.7
charset-normalizer==2.1.1
Django==3.2.13
django-cors-headers==4.0.0
djangorestframework==3.14.0
idna==3.4
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.15
gunicorn==20.1.0
